backend/requirements.txt
Flask==1.0.2
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
flask-swagger==0.2.14
flask-swagger-ui==3.20.9
requests==2.19.1

mysql-connector-python